Renamed com.teamscale.jacoco.util to com.teamscale.jacoco.agent.util
