+++ agent/src/main/java/com/teamscale/jacoco/agent/store/upload/UploadStoreBase.java
@@ -1,11 +1,10 @@
 package com.teamscale.jacoco.agent.store.upload;
 
-import com.teamscale.client.EReportFormat;
 import com.teamscale.jacoco.agent.store.IXmlStore;
 import com.teamscale.jacoco.agent.store.UploadStoreException;
 import com.teamscale.jacoco.agent.store.file.TimestampedFileStore;
-import com.teamscale.jacoco.util.Benchmark;
-import com.teamscale.jacoco.util.LoggingUtils;
+import com.teamscale.jacoco.agent.util.Benchmark;
+import com.teamscale.jacoco.agent.util.LoggingUtils;
 import okhttp3.HttpUrl;
 import okhttp3.ResponseBody;
 import org.conqat.lib.commons.filesystem.FileSystemUtils;
